+++ lib/src/poctal.dart
@@ -17,7 +17,7 @@ class OctalScanner extends NumberScanner {
     if (value == null) {
       return false;
     }
-    matches.add(neg ? -value : value);
+    addMatch(matches, neg ? -value : value);
     return true;
   }
 }